move 'allPass' tests to separate file
